pyproject.toml
[project]
name = "precondition-opt"
description = "Preconditioning optimizers."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [{name = "precondition authors", email="precondition-optimizers@google.com"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
]
keywords = []

# pip dependencies of the project
dependencies = [
  "chex",
  "flax",
  "optax",
  "praxis",
  "jax",
  "numpy"
]

# This is set automatically by flit using `precondition.__version__`
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/google-research/precondition"
repository = "https://github.com/google-research/precondition"
# Other: `documentation`, `changelog`

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install .[dev]`
dev = [
  "absl-py>=0.8.1",
  "pytest",
  "pytest-xdist",
  "pylint>=2.6.0",
  "pyink",
]

[tool.pyink]
# Formatting configuration to follow Google style-guide
line-length = 80
preview = true
pyink-indentation = 2
pyink-use-majority-quotes = true

[build-system]
requires = ["flit_core >=3.5,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "precondition"
